для удаления передаю не в теле запроса

+++ src/main/java/ru/yakov/demo/UserService/UserServiceImpl.java
@@ -21,7 +21,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(int id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        }
     }
 
     @Override
